New aproach
- legend list + life time stats
- with optional flag: list all legends with stats in separate embeds
